Implement the authors field as a textarea
The authors facet is implemented as a filterable text facet, but it does not have allowed values. In the legacy view it was rendered as a <ul> with the last <li> having an input.

The facet itself is registered as a type array of strings in publishing-api, so we need to map it from a string to an array, and vice versa, for writing and reading respectively. On the form we use the `author_tags` name and then do the conversion with the `author_tags` getters and setters in the model.

We decided to implement this field as a textarea in the design system. To do so, I tried having independent methods and virtual field names for the DS case (to aid with removal of code when we go live), but the `authors` value kept getting overridden. I eventually opted for a hidden field to mark whether we're on a legacy or a DS view. The hidden field `is_using_design_system_view` is always set to "true", because we're on a DS view, so the user has the "preview_design_system" permission. The edit pages required passing the virtual field from the controller.

Instructions to remove this code have been added to the "clean-up" card for removing legacy code.