Add access control tests for design system
The introduction of the "finders" resource was causing some incorrect behaviours regarding permissions. Because the finders index path was not protected, users with the DS permissions but no participating organisations were just "passed through" to `/finders`. They would have only seen the "All finders" title with a blank page, since their organisation does not have access to a finder, but the correct behaviour is to show the error page.

I replicated the redirect logic from the `PassthroughController` in the `FindersController`. I don't think there is much value in this `PassthroughController` anymore. It should be deleted when we clean up the feature toggle.

I also added a test to cover the redirect to /finder in the "Viewing the finders index" spec.