Change how we compute custom errors
The issue we're trying to fix here is the rendering of custom error messages/fields. These are currently only defined for licence transactions in the `en.yml` I18n file. The `LicenceTransaction model` is marked as having custom fields through the `custom_error_message_fields` method. This is again consumed in the `DocumentsController` - `document_error_messages` method, to figure out the errors - either custom or full messages. All this controller logic is bypassed if we are on the design system.

In the previous implementation, we tapped into this logic in the `ErrorSummaryComponent` to render the correct errors for Licences. But we missed the fact that the validation messages on the fields themselves were not being done properly (they're full message mixed with custom, giving the rather unsightly `Licence transaction industry Select an industry`).

This commit fixes both the `ErrorSummary` and the field errors, as well as opening up the possibility of adding custom errors for subsequent models without having to go through the model or controller. The model and controller code will stay in place for legacy, but once we remove the DS toggle code, we will be able to remove that complexity.

This is done by:
- moving the `get_text` method in the `ErrorsHelper` and reuse it for both fields and summary
- make the `get_text` method tap into the yml file directly
- we need to know what to search for in the yml file so we pass into `get_text`, the model "format" or document_type - which is how adding things to yml happens by convention
- this has caused many changes where the `errors_for` or `errors_for_input` were used, but it's simply passing that param through
- `errors_for` or `errors_for_input` - are mainly distinguished by the fact that they map to an array or an object with the `text` key in it, depending on what consumes them; for now we have to keep both
- `errors_for` is used practically across all the components under `FacetInputComponent` so we could extract it at the top in the `FacetInputComponent` initializer - but I think we can live with the repetition.