WIP
TODO - the ExportPresenter logic must be schema driven/generalized so that it works for all document types. Right now we only have the BusinessFinanceSupportSchemeExportPresenter class.

It is a bit problematic as the custom finders have other fields that are not included in the schema. For example the Business...scheme also exports continuation_link and will_continue_on