Change finder organisation form control from multiselect with search to repeatable single select with search
We require the order of the user's selection to be respected in the schema changes sent to Zendesk, as the order controls the order in which the organisations are listed on the finder content item when it is rendered on GOV.UK.

The select with search component in multiselect mode does not respect the order because it uses the native browser select element to store the selected values. Most browsers simply submit the selections in the order they appear in the select options list.

By using the add another component, we can allow the user to control the order in which the organisations will appear on GOV.UK.