Fix key generation
The output of `Date signed` as a field name, was `datesigned` as
a key. The `.underscore` wasn't doing what we might hope!

The reason this worked in the tests is that each word was
capitalised. The `underscore` method simply changes strings from
CamelCase to snake_case.

Whilst we're here, we have a number of values that use ampersands
and it would be nice to have these represented in the key if possible,
so now convert `&` to `and`.

And we also had a bug whereby any name field that had a comma in
it would have the comma preserved in the generated key. So we're
now stripping all invalid characters and leading/trailing underscores.