Refactor label/value/key derivation
The wrong abstractions were being used here, making the code
rather difficult to follow. Splitting out the 'extract value' and
'extract label' methods made it much easier to reason about, and
I've consolidated the behaviour that converts an arbitrary string
to a key-like string (with explicit 'separator' parameter of either
"-" or "_" accordingly).