Update documentation
- Testing:
Remove not very useful commands on `grep`-ing tests. This only filters the output, test suite still runs all tests. I think we've fixed the format dependency - bespoke tests are only for custom views and validations.

- Correct instructions on "format":
The line previously suggested that you should add the "format" to publishing api (which could mean the `filter.format`), but it's actually the document type that we use. Linked to the naming conventions readme. There was an attempt to decouple these notions, or make the usage more expressive, but it was closed - here https://github.com/alphagov/specialist-publisher/pull/3166

- Add link to instructions on running finder frontend