Do not show metadata for facets that are not allowed in the schema
Migrating the farming grants to a new schema left the old documents displaying odd metadata (it was rendering the entire array of values selected by the user). This ensures that, if the facet is not listed in the schema, we do not make an attempt to render its label in the summary.

We chose to omit the value altogether, rather than falling back to the old key.