Add a confirmation step to publishing_api rake tasks
I have added a confirmation step to the following rake tasks:
- publish_finders
- publish_finder

This is to ensure that the person running these rake tasks explicitly opts in to making these changes. I have also added a stub in the before block of each task in the publishing_api_spec file to prevent the rake task from waiting for actual user input during the test.